asyncComponent abstraction and folder scructure
