+++ src/components/asyncComponent/ChunkLoading.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Loading } from 'src/components/loading/Loading';
 
-const LoadableLoading = ({ pastDelay }) => {
+const ChunkLoading = ({ pastDelay }) => {
   // if loading has taken more than 200 milliseconds
   if (pastDelay) {
     return <Loading />;
@@ -10,4 +10,4 @@ const LoadableLoading = ({ pastDelay }) => {
   }
 };
 
-export default LoadableLoading;
+export default ChunkLoading;
