+++ src/components/asyncComponent/asyncComponent.js
@@ -1,0 +1,12 @@
+import loadable from 'react-loadable';
+import ChunkLoading from './ChunkLoading';
+
+
+export default function asyncComponent(importComponent) {
+  const LoadableComponent = loadable({
+    loader: importComponent,
+    loading: ChunkLoading
+  });
+
+  return LoadableComponent;
+}
