+++ src/components/asyncComponent/tests/ChunkLoading.test.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import LoadableLoading from '../LoadableLoading';
+import ChunkLoading from '../ChunkLoading';
 
 describe('LoadableLoading Component', () => {
 
   it('should return null if no props', () => {
-    const wrapper = shallow(<LoadableLoading />);
+    const wrapper = shallow(<ChunkLoading />);
 
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render loading if pastDelay is true', () => {
-    const wrapper = shallow(<LoadableLoading pastDelay={true}/>);
+    const wrapper = shallow(<ChunkLoading pastDelay={true}/>);
 
     expect(wrapper).toMatchSnapshot();
   });
