+++ src/pages/reports/summary/components/async/LineChart.scss
