+++ src/pages/templates/PublishedPage.js
@@ -11,7 +11,7 @@ import { templateById } from 'src/selectors/templates';
 
 // Components
 import Form from './components/Form';
-import Editor from './components/Editor';
+import Editor from './components/Editor'; // async
 import { Loading } from 'src/components';
 import { Page, Grid } from '@sparkpost/matchbox';
